Move classes in simulation module in separate modules
