+++ src/gsy_e/gsy_e_core/simulation/__init__.py
@@ -1,0 +1,24 @@
+"""
+Copyright 2018 Grid Singularity
+This file is part of Grid Singularity Exchange.
+This program is free software: you can redistribute it and/or modify
+it under the terms of the GNU General Public License as published by
+the Free Software Foundation, either version 3 of the License, or
+(at your option) any later version.
+This program is distributed in the hope that it will be useful,
+but WITHOUT ANY WARRANTY; without even the implied warranty of
+MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+GNU General Public License for more details.
+You should have received a copy of the GNU General Public License
+along with this program.  If not, see <http://www.gnu.org/licenses/>.
+"""
+from gsy_e.gsy_e_core.simulation.simulation import (Simulation, simulation_class_factory,
+                                                    run_simulation, CoefficientSimulation)
+
+
+__all__ = [
+    "Simulation",
+    "CoefficientSimulation",
+    "simulation_class_factory",
+    "run_simulation"
+]
